Add types.py
